Correct parsing for entire mca file
It is slow. I need to benchmark to find the bottleneck. Parallelism may
be needed in the future, if it takes more than a second to parse an mca
file. Part of the bottleneck is probably in IO, where I prettyprint
192M. Also my numeric casting is wildly inefficient. Attoparsec is
probably NOT the culprit. Anyway, just need to benchmark.