Glitchy printing working
I can now print blocks with positions relative to the section. However,
there are still some indexing errors. These are rare, only 1 in a 1000
or so. Overall, the composition of blocks looks right (reasonable
proportions of stone, ore, water, air, etc). For now, I am printing the
indexing errors as "glitch". They seem to have a peculiar affinity for
diorite.

There is also a memory issue to consider. The conversion to point cloud,
as currently implemented, increases the memory footprint over 64 fold. I
need a boxed datastructure with indices into a pallete.